+++ vm/label.h
@@ -100,7 +100,7 @@ class Label
     void bind(uint32_t offset) {
         assert(!bound());
         status_ = (offset << 1) | kBound;
-        assert(this->offset() == offset);
+        assert(static_cast<uint32_t>(this->offset()) == offset);
     }
 
     Label& operator =(const Label&) = delete;
@@ -181,7 +181,7 @@ class CodeLabelBase
     void bind(uint32_t offset) {
         assert(!bound());
         status_ = (offset << 1) | kBound;
-        assert(this->offset() == offset);
+        assert(static_cast<uint32_t>(this->offset()) == offset);
     }
 
   protected:
